+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Allow setting `space` or `padding` on account init to customise account size
 - Support for Anchor events
 - Allow using pubkeys as seeds
+- Support for importing from local files
 
 ### Changed